Moved SourceOrigin (again) into its own File
SourceOrigin and its implementations became nearly 100 lines. Having it in its own file keeps  Project.scala slimmer.
